Renamed dir `preloaders` to `loaders`
